restructuring and adding empty constructors
