+++ README.md
@@ -1,4 +1,11 @@
 # PostGRES_ext
 Spatial database course extension to PostGRES
 
+# Build
+Currently the build builds all the implementation files and builds postgres_ext.cpp file which includes all interfaces.
 
+```bash
+make clean
+make
+./postgres_ext
+```
