+++ include/RGPHalfSegment2D.h
@@ -12,7 +12,8 @@ struct RGPHalfSegment2D {
     RGPPoint2D dominantPoint;
 
     // Constructors
-
+    
+    RGPHalfSegment2D();
     RGPHalfSegment2D(RGPSegment2D s, RGPPoint2D dp);
     ~RGPHalfSegment2D();
