+++ include/RGPSegment2D.h
@@ -10,7 +10,8 @@ struct RGPSegment2D {
     RGPPoint2D point1, point2;
 
     // Constructors
-
+    
+    RGPSegment2D();
     RGPSegment2D(RGPPoint2D p1, RGPPoint2D p2);
     ~RGPSegment2D();
