+++ postgres_ext.cpp
@@ -1,0 +1,15 @@
+#include "iostream"
+
+// Including all headers to test
+#include "include/Region2D.h"
+#include "include/RGPHalfSegment2D.h"
+#include "include/RGPOperations.h"
+#include "include/RGPPoint2D.h"
+#include "include/RGPSegment2D.h"
+
+int main(void) {
+    Number n();
+    Number a(5, 7);
+    std::cout << "All includes successfull!" << std::endl;
+    return 0;
+}
