+++ src/RGPOperations.cpp
@@ -1,3 +1,3 @@
-#include "RGPOperations.h"
+#include "../include/RGPOperations.h"
 
 std::optional<RGPPoint2D> RGPOperations::intersectionOf(RGPSegment2D s1, RGPSegment2D s2) {}
