+++ src/RGPPoint2D.cpp
@@ -1,7 +1,8 @@
-#include "RGPPoint2D.h"
+#include "../include/RGPPoint2D.h"
 
 // Constructors
 
+RGPPoint2D::RGPPoint2D() {}
 RGPPoint2D::RGPPoint2D(Number x, Number y) {}
 RGPPoint2D::~RGPPoint2D() {}
