+++ src/RGPSegment2D.cpp
@@ -1,7 +1,8 @@
-#include "RGPSegment2D.h"
+#include "../include/RGPSegment2D.h"
 
 // Constructors
 
+RGPSegment2D::RGPSegment2D() {}
 RGPSegment2D::RGPSegment2D(RGPPoint2D p1, RGPPoint2D p2) {}
 RGPSegment2D::~RGPSegment2D() {}
